requirements.txt
Flask==2.3.3
stripe==5.5.0
python-dotenv==1.0.0
mysql-connector-python==9.2.0
bcrypt==4.2.1